requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
pytest>=7.4.0

